refactor(menu): 메뉴 데이터 최신화
