+++ backend/src/main/java/skkuchin/service/ServiceApplication.java
@@ -34,33 +34,15 @@ public class ServiceApplication {
 						  KeywordService keywordService,
 						  ReviewService reviewService) {
 		return args -> {
-
-			//userService.saveRole(new Role(null, "ROLE_USER"));
-			//userService.saveRole(new Role(null, "ROLE_ADMIN"));
 			userService.saveRole(Role.builder().name("ROLE_USER").build());
 			userService.saveRole(Role.builder().name("ROLE_ADMIN").build());
 
-			//admin 계정 생성
 			userService.saveAdmin(new UserDto.SignUpForm("스꾸친관리자", "admin", "12341234", "12341234",  16, Major.경영학과));
-
-			//test 계정 생성
 			userService.saveTestUser(new UserDto.SignUpForm("테스트", "test", "12341234", "12341234", 20, Major.건축학과));
-			//데이터 자동 주입
 
-			//String path = System.getProperty("user.dir") + "\\src\\main\\java\\skkuchin\\service\\data\\";
-
-			String path = System.getProperty("user.dir") + "/src/main/java/skkuchin/service/data/"; //Mac 공통 경로
-
-/*
-			String path = "C:\\Users\\sunny\\skkuchin_renewal2\\backend\\src\\main\\java\\skkuchin\\service\\data\\";
-*/
-
-			//tagService.insertData(path);
-			//placeService.insertData(path);
-			//imageService.insertData(path);
-			//menuService.insertData(path);
-			//keywordService.insertData(path);
-
+			//String path = System.getProperty("user.dir") + "\\src\\main\\java\\skkuchin\\service\\data\\"; // Window 공통 경로
+			String path = System.getProperty("user.dir") + "/src/main/java/skkuchin/service/data/"; // Mac 공통 경로
+//			String path = "C:\\Users\\sunny\\skkuchin_renewal2\\backend\\src\\main\\java\\skkuchin\\service\\data\\"; // 병준 경로
 
 			try {
 				tagService.insertData(path);
