+++ backend/src/main/java/skkuchin/service/config/chat/ChatConfig.java
@@ -1,32 +1,22 @@
 package skkuchin.service.config.chat;
 
 import lombok.RequiredArgsConstructor;
-import org.apache.tomcat.util.net.openssl.ciphers.Authentication;
-import org.springframework.amqp.rabbit.core.RabbitTemplate;
-import org.springframework.beans.factory.annotation.Required;
 import org.springframework.beans.factory.annotation.Value;
-import org.springframework.boot.autoconfigure.amqp.RabbitTemplateConfigurer;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.messaging.Message;
 import org.springframework.messaging.MessageChannel;
 import org.springframework.messaging.simp.config.ChannelRegistration;
 import org.springframework.messaging.simp.config.MessageBrokerRegistry;
-import org.springframework.messaging.simp.stomp.StompCommand;
-import org.springframework.messaging.simp.stomp.StompHeaderAccessor;
 import org.springframework.messaging.support.ChannelInterceptor;
-import org.springframework.messaging.support.MessageHeaderAccessor;
 import org.springframework.util.AntPathMatcher;
 import org.springframework.web.socket.config.annotation.EnableWebSocketMessageBroker;
 import org.springframework.web.socket.config.annotation.StompEndpointRegistry;
 import org.springframework.web.socket.config.annotation.WebSocketMessageBrokerConfigurer;
-import skkuchin.service.config.StompHandler;
 
 @Configuration
 @RequiredArgsConstructor
 @EnableWebSocketMessageBroker
 public class ChatConfig implements WebSocketMessageBrokerConfigurer {
-
-    private final StompHandler stompHandler;
 
     @Value("${rabbitmq.host}")
     private String host;
@@ -54,13 +44,6 @@ public class ChatConfig implements WebSocketMessageBrokerConfigurer {
                 .setRelayPort(port)
                 .setClientLogin(username)
                 .setClientPasscode(password);
-
-/*        registry.enableStompBrokerRelay("/topic")
-                .setRelayHost("localhost")
-                .setVirtualHost("/")
-                .setRelayPort(61613)
-                .setClientLogin("guest")
-                .setClientPasscode("guest");*/
     }
 
     @Override
@@ -68,8 +51,6 @@ public class ChatConfig implements WebSocketMessageBrokerConfigurer {
         registration.interceptors(new ChannelInterceptor() {
             @Override
             public Message<?> preSend(Message<?> message, MessageChannel channel) {
-                StompHeaderAccessor accessor =
-                        MessageHeaderAccessor.getAccessor(message, StompHeaderAccessor.class);
                 return message;
             }
         });
