--- backend/src/main/java/skkuchin/service/config/chat/Receiver.java
@@ -1,20 +1,0 @@
-package skkuchin.service.config.chat;
-
-import org.springframework.stereotype.Component;
-
-import java.util.concurrent.CountDownLatch;
-
-@Component
-public class Receiver {
-    private CountDownLatch latch = new CountDownLatch(1);
-
-    public void receiveMessage(String message) {
-        System.out.println("Received <" + message + ">");
-        latch.countDown();
-    }
-
-    public CountDownLatch getLatch() {
-        return latch;
-    }
-}
-
