+++ backend/src/main/java/skkuchin/service/config/chat/StompHandler.java
@@ -1,0 +1,28 @@
+package skkuchin.service.config.chat;
+
+import lombok.RequiredArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
+import org.springframework.messaging.Message;
+import org.springframework.messaging.MessageChannel;
+import org.springframework.messaging.simp.stomp.StompHeaderAccessor;
+import org.springframework.messaging.support.ChannelInterceptor;
+import org.springframework.stereotype.Component;
+
+@Slf4j
+@RequiredArgsConstructor
+@Component
+public class StompHandler implements ChannelInterceptor {
+
+    @Override
+    public Message<?> preSend(Message<?> message, MessageChannel channel) {
+        StompHeaderAccessor accessor = StompHeaderAccessor.wrap(message);
+        // websocket 연결시 헤더의 jwt token 검증
+     /*   if (StompCommand.CONNECT == accessor.getCommand()) {
+            jwtTokenProvider.validateToken(accessor.getFirstNativeHeader("token"));
+        }*/
+        return message;
+    }
+}
+
+
+
