--- backend/src/main/java/skkuchin/service/data/menu.json
@@ -1,109 +1,0 @@
-{
-  "menu": [
-    {
-      "placeId": 1,
-      "name": "기꾸초밥",
-      "price": 9000
-    },
-    {
-      "placeId": 1,
-      "name": "모듬초밥",
-      "price": 13000
-    },
-    {
-      "placeId": 1,
-      "name": "특모듬초밥",
-      "price": 18000
-    },
-    {
-      "placeId": 1,
-      "name": "광어초밥",
-      "price": 15000
-    },
-    {
-      "placeId": 1,
-      "name": "연어초밥",
-      "price": 15000
-    },
-    {
-      "placeId": 1,
-      "name": "생선회",
-      "price": 30000
-    },
-    {
-      "placeId": 1,
-      "name": "VIP초밥",
-      "price": 25000
-    },
-    {
-      "placeId": 2,
-      "name": "기꾸소바",
-      "price": 9000
-    },
-    {
-      "placeId": 2,
-      "name": "모듬소바",
-      "price": 13000
-    },
-    {
-      "placeId": 2,
-      "name": "특모듬소바",
-      "price": 18000
-    },
-    {
-      "placeId": 2,
-      "name": "광어소바",
-      "price": 15000
-    },
-    {
-      "placeId": 2,
-      "name": "연어소바",
-      "price": 15000
-    },
-    {
-      "placeId": 2,
-      "name": "생선소바",
-      "price": 30000
-    },
-    {
-      "placeId": 2,
-      "name": "VIP소바",
-      "price": 25000
-    },
-    {
-      "placeId": 3,
-      "name": "기꾸라멘",
-      "price": 9000
-    },
-    {
-      "placeId": 3,
-      "name": "모듬라멘",
-      "price": 13000
-    },
-    {
-      "placeId": 3,
-      "name": "특모듬라멘",
-      "price": 18000
-    },
-    {
-      "placeId": 3,
-      "name": "광어라멘",
-      "price": 15000
-    },
-    {
-      "placeId": 3,
-      "name": "연어라멘",
-      "price": 15000
-    },
-    {
-      "placeId": 3,
-      "name": "생선라멘",
-      "price": 30000
-    },
-    {
-      "placeId": 3,
-      "name": "VIP라멘",
-      "price": 25000
-    }
-  ]
-}
