+++ backend/src/main/resources/application.yml
@@ -1,20 +1,9 @@
 spring:
-
   #datasource:
   #  driver-class-name: com.mysql.cj.jdbc.Driver
   #  url: jdbc:mysql://localhost:3306/skkuchin3?characterEncoding=UTF-8&serverTimezone=UTC
   #  username: root
   #  password: 1234
- # datasource:
- #   driver-class-name: com.mysql.cj.jdbc.Driver
- #   url: jdbc:mysql://localhost:3306/service?characterEncoding=UTF-8&serverTimezone=UTC
- #   username: root
- #   password: Master00!@
-#  datasource:
- #   driver-class-name: com.mysql.cj.jdbc.Driver
- #   url: jdbc:mysql://localhost:3306/skkuchin1?characterEncoding=UTF-8&serverTimezone=UTC
- #   username: root
- #   password: 1234
 #  datasource:
 #    driver-class-name: com.mysql.cj.jdbc.Driver
 #    url: jdbc:mysql://localhost:3306/service?characterEncoding=UTF-8&serverTimezone=UTC
@@ -23,7 +12,6 @@ spring:
 
   jpa:
     hibernate:
-      #ddl-auto: update
       ddl-auto: create
       naming:
         physical-strategy: org.hibernate.boot.model.naming.CamelCaseToUnderscoresNamingStrategy
