--- backend/src/main/resources/templates/chat/room.html
@@ -1,93 +1,0 @@
-<!doctype html>
-<html lang="kr" xmlns:th="http://www.thymeleaf.org">
-<head>
-    <title>Websocket ChatRoom</title>
-    <!-- Required meta tags -->
-    <meta charset="utf-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1, maximum-scale=1, user-scalable=no">
-
-    <!-- Bootstrap CSS -->
-    <link rel="stylesheet" href="/webjars/bootstrap/4.3.1/dist/css/bootstrap.min.css">
-    <style>
-      [v-cloak] {
-          display: none;
-      }
-    </style>
-</head>
-<body>
-<div class="container" id="app" v-cloak>
-    <div class="row">
-        <div class="col-md-6">
-            <h3>{{roomName}}</h3>
-        </div>
-        <div class="col-md-6 text-right">
-            <a class="btn btn-primary btn-sm" href="/logout">로그아웃</a>
-        </div>
-    </div>
-    <div class="input-group">
-        <div class="input-group-prepend">
-            <h2 th:text="'Room No. ' + ${chatRoomId}"></h2>
-            <label class="input-group-text">내용</label>
-        </div>
-        <input type="text" class="form-control" v-model="message" v-on:keypress.enter="sendMessage('TALK')">
-        <div class="input-group-append">
-            <button class="btn btn-primary" type="button" @click="sendMessage('TALK')">보내기</button>
-        </div>
-    </div>
-    <ul class="list-group">
-        <li class="list-group-item" v-for="message in messages">
-            {{message.sender}} - {{message.message}}</a>
-        </li>
-    </ul>
-</div>
-<!-- JavaScript -->
-<script src="/webjars/vue/2.5.16/dist/vue.min.js"></script>
-<script src="/webjars/axios/0.17.1/dist/axios.min.js"></script>
-<script src="/webjars/sockjs-client/1.1.2/sockjs.min.js"></script>
-<script src="/webjars/stomp-websocket/2.3.3-1/stomp.min.js"></script>
-<script th:inline="javascript">
-        // websocket & stomp initialize
-        var sock = new SockJS("/ws/chat");
-        var ws = Stomp.over(sock);
-        var reconnect = 0;
-        const chatRoomId = [[${chatRoomId}]];
-        ws.heartbeat.outgoing = 0;
-        ws.heartbeat.incoming = 0;
-        // vue.js
-        var vm = new Vue({
-            el: '#app',
-            data: {
-                roomId: '',
-                roomName: '',
-                message: '',
-                messages: [],
-                token: ''
-            },
-            created() {
-
-                var _this = this;
-                axios.get('/chat/room').then(response => {
-                    _this.token = response.data.token;
-                    ws.connect({"token":_this.token}, function(frame) {
-                        ws.subscribe(`/exchange/chat.exchange/room.${chatRoomId}`, function(message) {
-                            var recv = JSON.parse(message.body);
-                            _this.recvMessage(recv);
-                        });
-                        _this.sendMessage('ENTER');
-                    }, function(error) {
-                        alert("서버 연결에 실패 하였습니다. 다시 접속해 주십시요.");
-                        location.href="/chat/room";
-                    });
-                });
-            },
-            methods: {
-                sendMessage: function(type) {
-                    ws.send(`/app/chat.message.${chatRoomId}`, {"token":this.token}, JSON.stringify({type:type, roomId:this.roomId, message:this.message}));
-                    this.message = '';
-                },
-
-            }
-        });
-    </script>
-</body>
-</html>
