+++ backend/src/main/resources/templates/chat/rrr.html
@@ -39,7 +39,7 @@
             message.innerHTML = "Logged in successfully";
             window.location.href = `${API_URL}/chat/room?chatRoomId=34296005-31d8-431e-948f-229b13994df2`
           } else {
-            message.innerHTML = "Authentication failed";
+            message.innerHTML = "OAS_30 failed";
           }
         } catch (err) {
           message.innerHTML = "Something went wrong when authenticating";
