+++ frontend/src/reducers/review.js
@@ -54,6 +54,8 @@ const reviewReducer = (state = initialState, action) => {
             return{
                 ...state
             }
+        default:
+            return state;
     }
 }
